new kivy update 09/05
