+++ libs/uix/baseclass/servicer_notification.py
@@ -1,0 +1,19 @@
+from kivy.core.window import Window
+from kivymd.uix.screen import MDScreen
+
+
+class ServiceNotification(MDScreen):
+    def __init__(self, **kwargs):
+        super(ServiceNotification, self).__init__(**kwargs)
+        Window.bind(on_keyboard=self.on_keyboard)
+
+    def on_keyboard(self, instance, key, scancode, codepoint, modifier):
+        if key == 27:  # Keycode for the back button on Android
+            self.on_back_button()
+            return True
+        return False
+
+    def on_back_button(self):
+        self.manager.push_replacement("servicer_dashboard", 'right')
+        screen = self.manager.get_screen('servicer_dashboard')
+        screen.ids.nav_drawer.set_state("close")
