+++ libs/uix/kv/servicer_add_service.kv
@@ -1,0 +1,131 @@
+
+#-------------------------Service-Slot-Adding-Page----------------------------
+<ServiceSlotAdding>
+    name:'service_slot_adding'
+    MDIconButton:
+        icon: "arrow-left"
+        pos_hint: {"center_y": .94}
+        user_font_size: "20sp"
+        theme_text_color: "Custom"
+        text_color: rgba(34, 34, 34, 255)
+        on_release:root.on_back_button()
+
+    MDIconButton:
+        icon: "dots-vertical"
+        pos_hint: {"center_x": .93, "center_y": .94}
+        user_font_size: "20sp"
+        theme_text_color: "Custom"
+        text_color: rgba(34, 34, 34, 255)
+
+    MDBoxLayout:
+        orientation: 'vertical'
+        size_hint_y: None
+        height: self.minimum_height
+#            MDTextField:
+#                id:slot_no
+#                hint_text: "Slot No"
+#        #                        required: True
+#                font_name: "Roboto"
+#                font_size: "14sp"
+#                pos_hint:{'center_x':.5}
+#                size_hint:.9,None
+#                height: "30dp"
+#                multiline: False
+#                color_mode: 'custom'
+#                line_color_focus: 0, 0, 0, 1
+#                text_color_focus:0,0,0,1
+#                hint_text_color:0,0,0,1
+#                hint_text_color_focus:0,0,0,.5
+#            MDTextField:
+#                id:add_date
+#                hint_text: "Date"
+#        #                        required: True
+#                font_name: "Roboto"
+#                font_size: "14sp"
+#                pos_hint:{'center_x':.5}
+#                size_hint:.9,None
+#                height: "30dp"
+#                multiline: False
+#                color_mode: 'custom'
+#                line_color_focus: 0, 0, 0, 1
+#                text_color_focus:0,0,0,1
+#                hint_text_color:0,0,0,1
+#                hint_text_color_focus:0,0,0,.5
+#                on_focus: root.show_date_picker(self)
+#            MDLabel:
+#                text:'Upload Slot Photos :'
+#                size_hint_y:.1
+#                font_size: "16sp"
+#            MDBoxLayout:
+#                orientation: 'horizontal'
+#                size_hint_y: None
+#                height: self.minimum_height
+#                padding:dp(20)
+#                spacing:dp(20)
+#
+#                MDFlatButton:
+#                    text: "Select File"
+#                    size_hint:None,None
+#                    height:dp(20)
+#                    width:dp(30)
+#                    on_release: root.file_manager_open()
+#                    pos_hint: {"center_x":.3,'center_y':.5}
+#                    background_color: 0, 0, 0, 0
+#                    font_name: "Roboto-Bold"
+#                    font_size: "12sp"
+#                    background_color: 0, 0, 0, 0
+#                    color: rgba(1, 1, 1, 255)
+#                    canvas.before:
+#                        Color:
+#                            rgb: rgba(255, 0, 0, 255)
+#                        Line:
+#                            width: 1.2
+#                            rounded_rectangle: self.x, self.y, self.width, self.height, 5, 5, 5, 5, 100
+#                MDFlatButton:
+#                    id: upload
+#                    text: "UPLOAD"
+#                    size_hint:None,None
+#                    height:dp(20)
+#                    width:dp(30)
+#                    pos_hint: {"center_x":.8,'center_y':.5}
+#                    background_color: 0, 0, 0, 0
+#                    font_name: "Roboto-Bold"
+#                    font_size: "12sp"
+#                    background_color: 0, 0, 0, 0
+#                    color: rgba(1, 1, 1, 255)
+#                    on_release: root.upload_file()
+#                    canvas.before:
+#                        Color:
+#                            rgb: rgba(255, 0, 0, 255)
+#                        Line:
+#                            width: 1.2
+#                            rounded_rectangle: self.x, self.y, self.width, self.height, 5, 5, 5, 5, 100
+#
+#            MDLabel:
+#                id: file_path
+#                text: "None selected"  # Initial text when no file is selected
+#        #                            font_name: "Roboto"
+#                italic: True
+#                color: 0.5, 0.5, 0.5, 1
+#                pos_hint: {"center_x": .5,'center_y':.4}
+#                size_hint_y: None
+#                height: "30dp"
+#
+#
+#            Button:
+#                id: solt_permission
+#                text: "GET PERMISSION"
+#                size_hint: .7,None
+#                height:dp(40)
+#                pos_hint: {"center_x": .5, "center_y": .34}
+#                background_color: 0, 0, 0, 0
+#                font_name: "Roboto-Bold"
+#                on_release:
+#                    root.solt_permission(self)
+#                canvas.before:
+#                    Color:
+#                        rgb: rgba(255, 0, 0, 255)
+#                    RoundedRectangle:
+#                        size: self.size
+#                        pos: self.pos
+#                        radius: [5]
