+++ libs/uix/kv/servicer_notification.kv
@@ -1,0 +1,35 @@
+
+#----------------------Service-Notification-Page-----------------------
+
+<ServiceNotification>
+    name:'service_notification'
+    MDFloatLayout:
+        md_bg_color: rgba(246, 250, 255, 255)
+        MDIconButton:
+            icon: "arrow-left"
+            pos_hint: {"center_y": .94}
+            user_font_size: "20sp"
+            theme_text_color: "Custom"
+            text_color: rgba(34, 34, 34, 255)
+            on_release:root.on_back_button()
+
+        MDIconButton:
+            icon: "dots-vertical"
+            pos_hint: {"center_x": .93, "center_y": .94}
+            user_font_size: "20sp"
+            theme_text_color: "Custom"
+            text_color: rgba(34, 34, 34, 255)
+
+        MDLabel:
+            text: "Notification"
+            font_name: "Roboto-Bold"
+            font_size: "20sp"
+            pos_hint: {"center_x": .56, "center_y": .88}
+            color: rgba(34, 34, 34, 255)
+        MDLabel:
+            text: "No current notifications"
+            font_size: "25sp"
+            halign:'center'
+            font_name: "Roboto"
+            pos_hint: {"center_x": .5, "center_y": .5}
+            color: rgba(94, 94, 94, 255)
