+++ libs/uix/kv/servicer_support.kv
@@ -1,0 +1,104 @@
+
+#----------------------------service-support-------------------------------------
+<ServiceSupport>
+    name:'service_support'
+    MDIconButton:
+        icon: "arrow-left"
+        pos_hint: {"center_y": .96}
+        user_font_size: "20sp"
+        theme_text_color: "Custom"
+        text_color: rgba(35, 35, 35, 255)
+        on_release:root.on_back_button()
+
+    MDIconButton:
+        icon: "dots-vertical"
+        pos_hint: {"center_x": 0.9, "center_y": 0.96}
+        user_font_size: "20sp"
+        theme_text_color: "Custom"
+        text_color: rgba(35, 35, 35, 255)
+
+    MDLabel:
+        text: "Support"
+        font_name: "Roboto-Bold"
+        font_size: "20sp"
+        pos_hint: {"center_x": .55, "center_y":0.90}
+        color: rgba(35, 35, 35, 255)
+
+    OneLineIconListItem:
+        text: "Contact Client Support"
+        pos_hint: {"center_x":0.55, "center_y":0.8}
+        on_release: root.show_customer_support_dialog()
+        IconLeftWidget:
+            icon: "face-agent"
+
+    OneLineIconListItem:
+        text: "Contact on-call Client Support"
+        pos_hint: {"center_x":0.55, "center_y":0.70}
+        on_release: root.show_doctor_dialog()
+        IconLeftWidget:
+            icon: "doctor"
+
+    MDLabel:
+        text: "Raise a ticket"
+        font_name: "Roboto-Bold"
+        font_size: "15sp"
+        pos_hint: {"center_x": .55, "center_y": .60}
+        color: rgba(35, 35, 35, 255)
+
+    TextInput:
+        id: issue_title
+        hint_text: "Issue Title"
+        font_name: "Roboto"
+        size_hint_y: 0.1
+        helper_text:'Required'
+        helper_text_mode:  'on_error'
+        required: True
+        pos_hint: {"center_x": .55, "center_y": .5}
+        background_color: 1, 1, 1, 0
+        foreground_color: rgba(0, 0, 59, 255)
+        cursor_color: rgba(35, 35, 35, 255)
+        font_size: "14sp"
+        cursor_width: "2sp"
+        multiline: False
+        write_tab: True
+        on_text_validate: root.clear_text_input()
+
+    TextInput:
+        id: issue_description
+        hint_text: "Issue description"
+        font_name: "Roboto"
+        size_hint_y: .1
+        helper_text:'Required'
+        helper_text_mode:  'on_error'
+        required: False
+        pos_hint: {"center_x": .55, "center_y": .4}
+        background_color: 1, 1, 1, 0
+        foreground_color: rgba(0, 0, 59, 255)
+        cursor_color: rgba(35, 35, 35, 255)
+        font_size: "14sp"
+        cursor_width: "2sp"
+        multiline: True
+        write_tab: True
+        on_text_validate: root.clear_text_input()
+
+    Button:
+        id: submit_ticket
+        text: "Submit Ticket"
+        size_hint: .5, .05
+        pos_hint: {"center_x": .3, "center_y": .2}
+        background_color: 0, 0, 0, 0
+        font_name: "Roboto-Bold"
+        canvas.before:
+            Color:
+                rgb: rgba(255, 0, 0, 255)
+            RoundedRectangle:
+                size: self.size
+                pos: self.pos
+                radius: [5]
+        on_release:
+            root.submit_ticket()
+            root.show_ticket_popup()
+
+
+<MDDialog>:
+    size_hint: 0.8, 0.3
